+++ src/main/kotlin/com/cognifide/gradle/aem/instance/CreateTask.kt
@@ -2,7 +2,7 @@ package com.cognifide.gradle.aem.instance
 
 import com.cognifide.gradle.aem.AemTask
 import com.cognifide.gradle.aem.deploy.SyncTask
-import com.cognifide.gradle.aem.internal.FileResolver
+import com.cognifide.gradle.aem.internal.file.FileResolver
 import groovy.lang.Closure
 import org.gradle.api.tasks.Internal
 import org.gradle.api.tasks.TaskAction
