+++ src/main/kotlin/com/cognifide/gradle/aem/pkg/PrepareTask.kt
@@ -3,7 +3,7 @@ package com.cognifide.gradle.aem.pkg
 import com.cognifide.gradle.aem.AemConfig
 import com.cognifide.gradle.aem.AemPackagePlugin
 import com.cognifide.gradle.aem.AemTask
-import com.cognifide.gradle.aem.internal.FileOperations
+import com.cognifide.gradle.aem.internal.file.FileOperations
 import org.apache.commons.io.FileUtils
 import org.gradle.api.DefaultTask
 import org.gradle.api.tasks.Nested
